refactor: move config analyzer to another file
